backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg[binary]==3.1.18
pydantic==2.6.0
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.26.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.6
